% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_llm.R
\name{instruction}
\alias{instruction}
\title{create translation instructions}
\usage{
instruction(from, to)
}
\arguments{
\item{from}{plain text language to translate from}

\item{to}{plain text language to translate into}
}
\value{
the translation instructions
}
\description{
create translation instructions
}
\examples{
paste0(instruction(from="English", to = "French"), "Let's have a look at mtcars\n```{r}\nsummary(mtcars)\n```")
}
