% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_llm.R
\name{ollama_api_get}
\alias{ollama_api_get}
\title{translate text through ollama API}
\usage{
ollama_api_get(
  model,
  model_prompt,
  url = "http://localhost:11434/api/generate"
)
}
\arguments{
\item{model}{the ollama model name to call}

\item{url}{the ollama API endpoint url}

\item{instruction}{the ollama instruction}
}
\value{
text generated by the model
}
\description{
translate text through ollama API
}
\examples{
ollama_api_get(model = "llama3.1:latest", 
              model_prompt = "Why is the moon round ? provide a short answer")
}
