% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_lang.R
\name{select_lang}
\alias{select_lang}
\title{Select Language}
\usage{
select_lang(lang)
}
\arguments{
\item{lang}{translation model to use like "Helsinki-NLP/opus-mt-en-es"}
}
\description{
for example "Helsinki-NLP/opus-mt-en-es"
}
