% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{translate an md file}
\usage{
translate(text)
}
\arguments{
\item{text}{Text from a md file}
}
\description{
\code{translate()} is best used in a map or apply statement
onto a parsed md file. the \code{lightparser} works seemlessly
with this function
}
\examples{
parsed <- lightparser::split_to_tbl("my.qmd")
parsed |>
  unnest(cols = text) |>
  mutate(text_es = map(text,translate))
}
