% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_llm.R
\name{translate_llm}
\alias{translate_llm}
\title{translate an md file via LLM}
\usage{
translate_llm(
  md,
  instruction = instruction(from = "English", to = "French"),
  url. = "http://localhost:11434/api/generate"
)
}
\arguments{
\item{md}{md file name}
}
\description{
\code{translate_llm()} will split the markdown file into chunks of around 100 lines
and make the llm proceed it through the instruction
}
\examples{
parsed <- lightparser::split_to_tbl("my.qmd")
parsed |>
  unnest(cols = text) |>
  mutate(text_es = map(text,translate))
}
